requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
PyJWT>=2.9,<3.0
httpx>=0.28,<1.0
